update default folder
